implementação do projeto de infra e api

+++ MovimentosManuais/src/MovimentosManuais.Application/DTOs/CriarMovimentoManualRequest.cs
@@ -1,8 +1,8 @@
 ﻿namespace MovimentosManuais.Application.DTOs;
 
 public sealed record CriarMovimentoManualRequest(
-    decimal Mes,
-    decimal Ano,
+    short Mes,
+    short Ano,
     string CodigoProduto,
     string CodigoCosif,
     decimal Valor,
