+++ MovimentosManuais/src/MovimentosManuais.Application/DTOs/EditarMovimentoManualRequest.cs
@@ -1,9 +1,9 @@
 ﻿namespace MovimentosManuais.Application.DTOs;
 
 public sealed record EditarMovimentoManualRequest(
-    decimal Mes,
-    decimal Ano,
-    decimal NumeroLancamento,
+    short Mes,
+    short Ano,
+    int NumeroLancamento,
     string CodigoProduto,
     string CodigoCosif,
     decimal Valor,
