+++ MovimentosManuais/src/MovimentosManuais.Application/DTOs/ExcluirMovimentoManualRequest.cs
@@ -1,6 +1,6 @@
 ﻿namespace MovimentosManuais.Application.DTOs;
 
 public sealed record ExcluirMovimentoManualRequest(
-    decimal Mes,
-    decimal Ano,
-    decimal NumeroLancamento);
+    short Mes,
+    short Ano,
+    int NumeroLancamento);
